refactor: Move exported types to `types` directory and update imports to `import type` where relevant
