+++ src/commands/remindMe.ts
@@ -3,7 +3,7 @@ import {
   ChatInputCommandInteraction,
   Message,
 } from "discord.js";
-import { Command } from "../interfaces/command.interface.ts";
+import type { Command } from "../interfaces/command.interface.ts";
 import {
   getIntegerInput,
   getStringInputFromRest,
