+++ src/commands/rps.ts
@@ -3,7 +3,7 @@ import {
   ChatInputCommandInteraction,
   Message,
 } from "discord.js";
-import { Command } from "../interfaces/command.interface.ts";
+import type { Command } from "../interfaces/command.interface.ts";
 import { getStringInput } from "../utilities/commandContext.ts";
 import { sendMessage } from "../utilities/sendMessage.ts";
