+++ src/commands/userInfo.ts
@@ -4,7 +4,7 @@ import {
   EmbedBuilder,
   Message,
 } from "discord.js";
-import { Command } from "../interfaces/command.interface.ts";
+import type { Command } from "../interfaces/command.interface.ts";
 import { filterUserId } from "../utilities/filterUserId.ts";
 import { sendMessage } from "../utilities/sendMessage.ts";
