+++ src/interfaces/client.interface.ts
@@ -1,5 +1,5 @@
-import { Client, ClientOptions } from "discord.js";
-import { Command } from "./command.interface.ts";
+import type { Client, ClientOptions } from "discord.js";
+import type { Command } from "./command.interface.ts";
 
 export interface AdditionalOptions extends ClientOptions {
   prefix: string;
@@ -8,8 +8,3 @@ export interface AdditionalOptions extends ClientOptions {
 export interface CustomClientProperties extends Client {
   commands: Map<string, Command>;
 }
-
-export type CustomClientOptions = ClientOptions & AdditionalOptions;
-
-export type CustomClient = Client &
-  CustomClientProperties & { options: CustomClientOptions };
