+++ src/types/sendMessage.type.ts
@@ -1,0 +1,14 @@
+import type {
+  InteractionReplyOptions,
+  MessageCreateOptions,
+  MessagePayload,
+} from "discord.js";
+
+export type MessageContent = string | MessagePayload | MessageCreateOptions;
+
+export type InteractionContent =
+  | string
+  | MessagePayload
+  | InteractionReplyOptions;
+
+export type dualContentsType = MessageContent | InteractionContent;
